shavtsak-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
python-dotenv
pydantic
passlib[bcrypt]
httpx
